Holy Refactor
After 1 year and 3 months, returning to this project required a lot.
The biggest change was to switch from MySQL to Postgres.
Then TypeORM updates, as there was lots of deprecation.
Then using Docker for the application itself.
Setting up development Docker container that watches for file changes.
Removing pathing for now from tsconfig.
42 files in this, which is wild.