Yet Another Large Refactor
Now, mod based commands have an extensible class that forces the mod
check such that the `exec` function doesn't have to.

In turn that means regular commands now extend UserCommand instead of
implementing Command.

Made it so points are not awarded to users in chat when channel is not
live.

Added the ability to create response commands via a `cmds` table. These
commands should only be information based, as they can only respond with
text.